# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"UserRole": {
				"type": "string",
				"enum": [
					"admin",
					"user"
				]
			},
			"IUserPayload": {
				"properties": {
					"user_id": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/UserRole"
					}
				},
				"required": [
					"user_id",
					"username",
					"email",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserLoginResponse": {
				"properties": {
					"user": {
						"$ref": "#/components/schemas/IUserPayload"
					},
					"token": {
						"type": "string"
					},
					"refreshToken": {
						"type": "string"
					}
				},
				"required": [
					"user",
					"token",
					"refreshToken"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "api",
		"version": "1.0.0",
		"description": "joel",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "joel"
		}
	},
	"servers": [
		{
			"url": "/"
		}
	],
	"paths": {
...
		"/auth/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Informations sur l'utilisateur et tokens d'authentification",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IUserLoginResponse"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"description": "Enregistre un nouvel utilisateur",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Données d'enregistrement de l'utilisateur",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									},
									"username": {
										"type": "string"
									}
								},
								"required": [
									"password",
									"email",
									"username"
								],
								"type": "object",
								"description": "Données d'enregistrement de l'utilisateur"
							}
						}
					}
				}
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Login successful",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IUserLoginResponse"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"description": "Authentifie un utilisateur",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Données de connexion de l'utilisateur",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"password",
									"email"
								],
								"type": "object",
								"description": "Données de connexion de l'utilisateur"
							}
						}
					}
				}
			}
		},
		"/auth/renew": {
			"post": {
				"operationId": "RefreshToken",
				"responses": {
					"200": {
						"description": "Nouveau token d'accès",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"token": {
											"type": "string"
										}
									},
									"required": [
										"token"
									],
									"type": "object",
									"nullable": true
								}
							}
						}
					}
				},
				"description": "Rafraîchit un token d'accès",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Token de rafraîchissement",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"refreshToken": {
										"type": "string"
									}
								},
								"required": [
									"refreshToken"
								],
								"type": "object",
								"description": "Token de rafraîchissement"
							}
						}
					}
				}
			}
		},
		"/auth/verify": {
			"get": {
				"operationId": "VerifyToken",
				"responses": {
					"200": {
						"description": "Informations sur l'utilisateur décodées du token",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUserPayload"
								}
							}
						}
					}
				},
				"description": "Vérifie un token JWT",
				"tags": [
					"Authentication"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/auth/logout": {
			"post": {
				"operationId": "Logout",
				"responses": {
					"200": {
						"description": "Statut de la déconnexion",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Déconnecte un utilisateur (révoque son refresh token)",
				"tags": [
					"Authentication"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Token de rafraîchissement à révoquer",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"refreshToken": {
										"type": "string"
									}
								},
								"required": [
									"refreshToken"
								],
								"type": "object",
								"description": "Token de rafraîchissement à révoquer"
							}
						}
					}
				}
			}
		},
...
	}
}